requirements copy.txt
fastapi==0.115.5
uvicorn==0.32.1
pandas>=2.1.3
anthropic==0.40.0
python-multipart==0.0.19
pdfplumber==0.11.4
pdf2image==1.17.0
openpyxl==3.1.5
python-dotenv==1.0.1
numpy>=1.26.4
supabase==2.3.5
psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
faker==22.6.0
aiohttp==3.9.3
async-timeout==4.0.3
beautifulsoup4==4.12.3
google-generativeai==0.3.2
pydantic==2.6.1